% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ageCut}
\alias{ageCut}
\title{Age cut}
\usage{
ageCut(x, lower = 0, upper, by = 10, sep = "-", above.char = "+")
}
\arguments{
\item{x}{A vector of ages}

\item{lower}{Lower bound for age bins}

\item{upper}{Upper bound for ang bins}

\item{by}{Size of age bins}

\item{sep}{Separator character for bin labels}

\item{above.char}{Character indicating top bin}
}
\description{
Define cuts in age vector
}
