% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{barAndLinePal}
\alias{barAndLinePal}
\title{Palettes for bar and line plots}
\usage{
barAndLinePal(fill)
}
\arguments{
\item{fill}{Name of color or fill variable (string)}
}
\value{
A character vector specifying a palette
}
\description{
Define color palette for each variable used in bar and line plots
}
