% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildLinePlot}
\alias{buildLinePlot}
\title{Build a line plot}
\usage{
buildLinePlot(df, x, y, fill = "none", facet = "none", title = NULL,
  facetScales = "fixed", yLimits = c(NA, NA), yScales = "zero",
  scaleLabels = waiver())
}
\arguments{
\item{df}{A data frame with plot data}

\item{x}{Name of x-variable (string)}

\item{y}{Name of y-variable (string)}

\item{fill}{Name of fill variable (string)}

\item{facet}{Name of facet variable (string)}

\item{title}{Title of plot (string)}

\item{facetScales}{Scale specification for subplots (fixed, free_y, free_x, free).
This value is passed to the \code{\link[lemon]{facet_rep_wrap}} function.}

\item{yLimits}{A vector specifying y-axis limits for use with
\code{yScales = "manual"}. Use \code{NA} to leave a limit unspecified.}

\item{yScales}{A string specifying y-axis scale type. Possible values are
"zero" for a zero lower limit and upper limit fit to data, "manual" for a manual
specification given by yLimits, or NULL for limits fit to data}

\item{scaleLabels}{Scale labels for y-variable (e.g., waiver(), scales::percent).
This value is passed to the \code{\link[ggplot2]{scale_y_continuous}} function.}
}
\value{
A ggplot2 plot object
}
\description{
This function builds a line plot given user specifications using ggplot2.
}
\details{
The plot labels are based on variable names. The
\code{\link[snakecase]{to_any_case}} function is used to convert variable
names from the data frame in camel case to sentence case.

The y-variable values are divided by one thousand or one million when
appropriate to simplify axis labels. Scaling is performed using the
\code{\link{scaleVariable}} function.

The y-axis labels are repeated for each
subplot using the \code{\link[lemon]{facet_rep_wrap}} function.

The \code{\link{barAndLinePal}} function is used to map the fill variable
to a fill color.

If the maximum length of x-variable labels are greater than or equal to
four, then the labels are rotated 45 degrees.
}
\seealso{
Other plot functions: \code{\link{buildBarPlot}}
}
\concept{plot functions}
