% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildMapPlot}
\alias{buildMapPlot}
\title{Build map plot}
\usage{
buildMapPlot(df, y, facet = "none", scaleLabels = waiver(),
  title = NULL)
}
\arguments{
\item{df}{A data frame with plot data}

\item{y}{Name of fill variable (string)}

\item{facet}{Name of facet variable (string)}

\item{scaleLabels}{Scale labels for y-variable (e.g., waiver(), scales::percent).
This value is passed to the \code{\link[viridis]{scale_fill_viridis}} function.}

\item{title}{Title of plot (string)}
}
\value{
A ggplot2 plot object
}
\description{
This function builds a choropleth plot given user specifications using ggplot2.
}
\details{
The plot labels are based on variable names. The
\code{\link[snakecase]{to_any_case}} function is used to convert variable
names from the data frame in camel case to sentence case.

If the fill variable is "permits" or "customers", then the data is log
transformed automatically.

The y-axis labels are repeated for each
subplot using the \code{\link[lemon]{facet_rep_wrap}} function.
}
