% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildRadialSetsPlot}
\alias{buildRadialSetsPlot}
\title{Build radial sets plot}
\usage{
buildRadialSetsPlot(sharedCustomers, degreeCount, returnPlot = F,
  linkThickness = "percent", focusGroup = "none",
  sectorLabelFontSize = 1.5, axisLabelFontSize = 1,
  maxLinkThickness = 15, countScale = 1/1000)
}
\description{
Build radial sets plot
}
