% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calcAvgValue}
\alias{calcAvgValue}
\title{Average participants by group}
\usage{
calcAvgValue(df, var, startYear, endYear, groupVars = NULL)
}
\arguments{
\item{df}{A data frame}

\item{startYear}{An integer}

\item{endYear}{An integer}

\item{groupVars}{A vector}
}
\description{
Calculate average value by group
}
