% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calcChurn}
\alias{calcChurn}
\title{Calculate churn}
\usage{
calcChurn(df, groupVars = NULL, deathThreshold = 5)
}
\arguments{
\item{df}{A data frame}

\item{groupVars}{A character vector of variable names to group by}

\item{deathThreshold}{A numeric value indicating how many years
must elapse before an individual is considered churned}
}
\value{
A dataframe with columns for grouping variables and columns
churnedCustomers, retainedCustomers, totalCustomers, churnRate, and
retentionRate
}
\description{
Calculate churn using variable window size
}
\examples{
# Demo data: Churn for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "csv",
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcChurn(c("itemType"))
\dontrun{
# Database connection. Suggest using keyring package to avoid hardcoding
# passwords
myConn <- DBI::dbConnect(odbc::odbc(),
  dsn = "HuntFishApp", # Your datasource name
  uid = keyring::key_get("HuntFishAppUID"), # Your username
  pwd = keyring::key_get("HuntFishAppPWD")
) # Your password

# SQL Backend: Churn for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "sql",
  conn = myConn,
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcChurn(c("itemType"))
}

}
\seealso{
Other analysis functions: \code{\link{calcGenderProportion}},
  \code{\link{calcParticipation}},
  \code{\link{calcRecruitment}},
  \code{\link{countCustomers}}, \code{\link{countItems}},
  \code{\link{itemGroupCount}}, \code{\link{sumRevenue}}
}
\concept{analysis functions}
