% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calcGenderProportion}
\alias{calcGenderProportion}
\title{Calculate gender proportion}
\usage{
calcGenderProportion(df, groupVars = NULL)
}
\arguments{
\item{df}{A data frame with a column named customerUID}

\item{groupVars}{A character vector of variable names to group by}
}
\value{
A dataframe with columns for grouping variables and a column named
\code{genderProportion} for gender proportion. Data frame is passed through
\code{\link{prettyData}} function.
}
\description{
Calculate gender proportion by group. This function collects data from
the database.
}
\examples{
# Demo data: Gender proportions for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "csv",
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcGenderProportion(c("itemYear", "itemType"))
\dontrun{
# Database connection. Suggest using keyring package to avoid hardcoding
# passwords
myConn <- DBI::dbConnect(odbc::odbc(),
  dsn = "HuntFishApp", # Your datasource name
  uid = keyring::key_get("HuntFishAppUID"), # Your username
  pwd = keyring::key_get("HuntFishAppPWD")
) # Your password

# SQL Backend: Gender proportions for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "sql",
  conn = myConn,
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcGenderProportion(c("itemYear", "itemType"))
}

}
\seealso{
Other analysis functions: \code{\link{calcChurn}},
  \code{\link{calcParticipation}},
  \code{\link{calcRecruitment}},
  \code{\link{countCustomers}}, \code{\link{countItems}},
  \code{\link{itemGroupCount}}, \code{\link{sumRevenue}}
}
\concept{analysis functions}
