% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calcRecruitment}
\alias{calcRecruitment}
\title{Calculate recruitment}
\usage{
calcRecruitment(df, groupVars = NULL, birthThreshold = 5)
}
\arguments{
\item{df}{A data frame}

\item{groupVars}{A character vector of variable names to group by}

\item{birthThreshold}{A numeric value indicating how many years
without purchase before an individual is considered a recruit}
}
\value{
A dataframe with columns for grouping variables and columns
for totalCustomers, retainedCustomers, recruitedCustomers, and recruitementRate
}
\description{
Calculate recruitment using a variable window size
}
\examples{
# Demo data: Recruitment for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "csv",
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcRecruitment(c("itemType"))
\dontrun{
# Database connection. Suggest using keyring package to avoid hardcoding
# passwords
myConn <- DBI::dbConnect(odbc::odbc(),
  dsn = "HuntFishApp", # Your datasource name
  uid = keyring::key_get("HuntFishAppUID"), # Your username
  pwd = keyring::key_get("HuntFishAppPWD")
) # Your password

# SQL Backend: Recruitment for customers purchasing a fishing
# license between 2010 and 2017
filterData(
  dataSource = "sql",
  conn = myConn,
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  calcRecruitment(c("itemType"))
}

}
\seealso{
Other analysis functions: \code{\link{calcChurn}},
  \code{\link{calcGenderProportion}},
  \code{\link{calcParticipation}},
  \code{\link{countCustomers}}, \code{\link{countItems}},
  \code{\link{itemGroupCount}}, \code{\link{sumRevenue}}
}
\concept{analysis functions}
