% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plot_height.R
\name{calcualtePlotHeight}
\alias{calcualtePlotHeight}
\title{Calculate plot height}
\usage{
calcualtePlotHeight(plotData, pageWidth, facetVars)
}
\arguments{
\item{plotData}{A dataframe of data used in plot}

\item{pageWidth}{Width of page in pixels}

\item{facetVars}{A character vector with names of facet variables}
}
\description{
Calculate plot height for different page widths and number of subplots
}
