% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_dplyr.R
\name{conditional_filter}
\alias{conditional_filter}
\title{Conditional filter function}
\usage{
conditional_filter(df1, cond, varName, varCond)
}
\arguments{
\item{df1}{A dataframe}

\item{cond}{A logical value indicating if filter should be applied}

\item{varName}{Name of variable to filter on}

\item{varCond}{A single value for equality filter or a vector for
is member filter}
}
\description{
Apply a filter if condition is met. This function is a simple wrapper for
\code{\link[dplyr]{filter}}
}
\examples{
data(starwars, package = "dplyr")
starwars \%>\%
  conditional_filter(TRUE, "species", "Human")

starwars \%>\%
  conditional_filter(TRUE, "species", c("Human", "Droid"))
}
\seealso{
Other conditional dplyr: \code{\link{conditional_mutate}},
  \code{\link{conditional_range_filter}},
  \code{\link{conditional_select}}
}
\concept{conditional dplyr}
