% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_dplyr.R
\name{conditional_range_filter}
\alias{conditional_range_filter}
\title{Range filter function}
\usage{
conditional_range_filter(df1, cond, varName, varCond)
}
\arguments{
\item{df1}{A dataframe}

\item{cond}{A logical value indicating if filter should be applied}

\item{varName}{Name of variable to filter on}

\item{varCond}{A vector with lower and upper bound on variable. Use NA
values to perform a one sided filter.}
}
\description{
Apply a filter on range of variable if condition is met This function is a simple wrapper for
\code{\link[dplyr]{filter}}
}
\examples{
data(starwars, package = "dplyr")
starwars \%>\%
  conditional_range_filter(TRUE, "height", c(160, 180))

starwars \%>\%
  conditional_range_filter(TRUE, "height", c(180, NA))
}
\seealso{
Other conditional dplyr: \code{\link{conditional_filter}},
  \code{\link{conditional_mutate}},
  \code{\link{conditional_select}}
}
\concept{conditional dplyr}
