% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_dplyr.R
\name{conditional_select}
\alias{conditional_select}
\title{Conditional select function}
\usage{
conditional_select(df1, cond, user_expr)
}
\arguments{
\item{df1}{A dataframe}

\item{cond}{A logical value indicating if select should be performed}

\item{user_expr}{Select expression}
}
\description{
Perform select if condition is met. This function is a simple wrapper for
\code{\link[dplyr]{select}}
}
\examples{
mtcars \%>\%
  dplyr::as_tibble() \%>\%
  conditional_select(
    exists("cyl", where = .),
    user_expr = -cyl
  )
}
\seealso{
Other conditional dplyr: \code{\link{conditional_filter}},
  \code{\link{conditional_mutate}},
  \code{\link{conditional_range_filter}}
}
\concept{conditional dplyr}
