% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyCentroids.R
\docType{data}
\name{countyCentroids}
\alias{countyCentroids}
\title{County centroids}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3084 rows and 4 columns.}
\usage{
countyCentroids
}
\description{
A dataset of centroids of counties by state.
The county centroids are used for calculating the nearest county
during tooltip interactions.
}
\seealso{
Other map data: \code{\link{countyMaps}},
  \code{\link{stateCentroids}}, \code{\link{stateMap}}
}
\concept{map data}
\keyword{datasets}
