% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyMaps.R
\docType{data}
\name{countyMaps}
\alias{countyMaps}
\title{County maps}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87949 rows and 9 columns.}
\usage{
countyMaps
}
\description{
A dataset for plotting a map of counties by state.
The map data is from \code{\link[ggplot2]{map_data}}. The data has been
modified to include additional variables for county centroids (\code{longCent},
\code{latCent}). The county centroids are used for calculating the nearest county
during tooltip interactions.
}
\seealso{
Other map data: \code{\link{countyCentroids}},
  \code{\link{stateCentroids}}, \code{\link{stateMap}}
}
\concept{map data}
\keyword{datasets}
