% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{createLineTooltip}
\alias{createLineTooltip}
\title{Tooltip for line plots}
\usage{
createLineTooltip(plotData, hover)
}
\arguments{
\item{plotData}{A data frame with data used in plot}

\item{hover}{A named list created by \code{\link[shiny]{plotOutput}}}
}
\value{
A tooltip created as a panel using \code{\link[shiny]{wellPanel}}
}
\description{
Buid tooltip for line plots
}
