% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{itemGroupCount}
\alias{itemGroupCount}
\title{Count customers by permit groups}
\usage{
itemGroupCount(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Count customers who bought each combination of permits and stamps
}
\examples{
# Demo data: Count number of customers by combination of items purchased
filterData(
  dataSource = "csv",
  activeFilters = list(
    itemType = c("Fish", "Deer", "Spring Turkey"),
    itemYear = c(2018, 2018)
  )
) \%>\%
  itemGroupCount()
\dontrun{
# Database connection. Suggest using keyring package to avoid hardcoding
# passwords
myConn <- DBI::dbConnect(odbc::odbc(),
  dsn = "HuntFishApp", # Your datasource name
  uid = keyring::key_get("HuntFishAppUID"), # Your username
  pwd = keyring::key_get("HuntFishAppPWD")
) # Your password

# SQL Backend: Count number of customers by combination of items purchased
filterData(
  dataSource = "sql",
  conn = myConn,
  activeFilters = list(
    itemType = c("Fish", "Deer", "Spring Turkey"),
    itemYear = c(2018, 2018)
  )
) \%>\%
  itemGroupCount()
}

}
\seealso{
Other analysis functions: \code{\link{calcChurn}},
  \code{\link{calcGenderProportion}},
  \code{\link{calcParticipation}},
  \code{\link{calcRecruitment}},
  \code{\link{countCustomers}}, \code{\link{countItems}},
  \code{\link{sumRevenue}}
}
\concept{analysis functions}
