% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauseable.R
\name{pause}
\alias{pause}
\title{Pause a reactive}
\usage{
pause(pausableReactive, pause = TRUE)
}
\arguments{
\item{pausableReactive}{A pauseable reactive}

\item{pause}{A logical indicative if paused}
}
\description{
Pause a pauseable reactive. Code from jcheng5
https://gist.github.com/jcheng5/cc6647a4971b8125c1bb
}
