% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauseable.R
\name{pauseableReactive}
\alias{pauseableReactive}
\title{Pauseable reactive}
\usage{
pauseableReactive(x, env = parent.frame(), quoted = FALSE,
  priority = 0, domain = shiny::getDefaultReactiveDomain(),
  paused = TRUE)
}
\arguments{
\item{x}{For \code{reactive}, an expression (quoted or unquoted). For
\code{is.reactive}, an object to test.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression.}

\item{quoted}{Is the expression quoted? By default, this is \code{FALSE}.
This is useful when you want to use an expression that is stored in a
variable; to do so, it must be quoted with \code{quote()}.}

\item{domain}{See \link{domains}.}

\item{paused}{Indicates the starting state of the pauseable reactive.}
}
\description{
Create a pauseable reactive. Code from jcheng5
https://gist.github.com/jcheng5/cc6647a4971b8125c1bb
}
