% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{radialSetsData}
\alias{radialSetsData}
\title{Calculate edge matrix for Radial Sets plot}
\usage{
radialSetsData(sharedCustomers, degreeCount, linkThickness = "percent",
  focusGroup = "none", countScale = 1/1000)
}
\description{
Calculate edge matrix for Radial Sets plot
}
