% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_API_functions.R
\name{runCensusAPIQueries}
\alias{runCensusAPIQueries}
\title{Census API function}
\usage{
runCensusAPIQueries(data_url, key, vars, region, numeric = TRUE)
}
\arguments{
\item{data_url}{The url root of the api including the '?'}

\item{key}{API key}

\item{vars}{A character vector of variable names}

\item{region}{A region specification containing a 'for:' and possibly an 'in:'}
}
\description{
Query data from US Census Bureau API and collect result into dplyr data frame
}
\details{
For information on the US Census API see \url{https://www.census.gov/developers/}.

This function is a modified version of the code published here:
\url{http://rstudio-pubs-static.s3.amazonaws.com/19337_2e7f827190514c569ea136db788ce850.html}
}
\examples{
\donttest{
runCensusAPIQueries(
  data_url = "https://api.census.gov/data/2017/acs/acs5?",
  key = rstudioapi::askForSecret("censusKey"),
  vars = paste0("B01001_", str_pad(2:49, 3, pad = "0"), "E"),
  region = "for=county:*&in=state:31"
)
}

}
\seealso{
Other census functions: \code{\link{queryCensusAPI}},
  \code{\link{statePop}}, \code{\link{vecToChunk}}
}
\concept{census functions}
