% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{scaleVariable}
\alias{scaleVariable}
\title{Round axis labels}
\usage{
scaleVariable(df, y, groupVars = NULL)
}
\arguments{
\item{df}{A data frame}

\item{y}{Name of y-variable to scale}

\item{groupVars}{Variables to group by}
}
\description{
Scale axis labels
}
