% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateCentroids.R
\docType{data}
\name{stateCentroids}
\alias{stateCentroids}
\title{State centroids}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 3 columns.}
\usage{
stateCentroids
}
\description{
A dataset of centroids for continental US states.
The state centroids are used for calculating the nearest state
during tooltip interactions.
}
\seealso{
Other map data: \code{\link{countyCentroids}},
  \code{\link{countyMaps}}, \code{\link{stateMap}}
}
\concept{map data}
\keyword{datasets}
