% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateMap.R
\docType{data}
\name{stateMap}
\alias{stateMap}
\title{State map}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15537 rows and 9 columns.}
\usage{
stateMap
}
\description{
A dataset for plotting a map of continental US states.
The map data is from \code{\link[ggplot2]{map_data}}. The data has been
modified to include additional variables for state centroids (\code{longCent},
\code{latCent}). The state centroids are used for calculating the nearest state
during tooltip interactions.
}
\seealso{
Other map data: \code{\link{countyCentroids}},
  \code{\link{countyMaps}}, \code{\link{stateCentroids}}
}
\concept{map data}
\keyword{datasets}
