% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{sumRevenue}
\alias{sumRevenue}
\title{Sum revenue by group}
\usage{
sumRevenue(df, groupVars = NULL)
}
\arguments{
\item{df}{A data frame}

\item{groupVars}{A character vector of variable names to group by}
}
\value{
A dataframe with columns for grouping variables and a column named
revenue for total revenue
}
\description{
Calculate total revenue by group. Total revenue is the sum price variable.
}
\examples{
# Demo data: Calculate revenue each year from fishing licenses between 2010
# and 2017
filterData(
  dataSource = "csv",
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  sumRevenue(c("itemYear", "itemType"))
\dontrun{
# Database connection. Suggest using keyring package to avoid hardcoding
# passwords
myConn <- DBI::dbConnect(odbc::odbc(),
  dsn = "HuntFishApp", # Your datasource name
  uid = keyring::key_get("HuntFishAppUID"), # Your username
  pwd = keyring::key_get("HuntFishAppPWD")
) # Your password

# SQL Backend: Calculate revenue each year from fishing licenses between 2010
# and 2017
filterData(
  dataSource = "sql",
  conn = myConn,
  activeFilters = list(itemType = "Fish", itemYear = c(2010, 2017))
) \%>\%
  sumRevenue(c("itemYear", "itemType"))
}

}
\seealso{
Other analysis functions: \code{\link{calcChurn}},
  \code{\link{calcGenderProportion}},
  \code{\link{calcParticipation}},
  \code{\link{calcRecruitment}},
  \code{\link{countCustomers}}, \code{\link{countItems}},
  \code{\link{itemGroupCount}}
}
\concept{analysis functions}
