% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_any_case_wrap.R
\name{to_any_case_wrap}
\alias{to_any_case_wrap}
\title{Wrap to_any_case function}
\usage{
to_any_case_wrap(string, case)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Wrapper for \code{\link[snakecase]{to_any_case}} function that returns NULL
if input string is NULL
}
