% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_API_functions.R
\name{vecToChunk}
\alias{vecToChunk}
\title{Split vector into chunks}
\usage{
vecToChunk(x, max = 50)
}
\arguments{
\item{x}{A vector}

\item{max}{Size of chunks}
}
\description{
Helper function to split Census API queries into chunks
}
\seealso{
Other census functions: \code{\link{queryCensusAPI}},
  \code{\link{runCensusAPIQueries}}, \code{\link{statePop}}
}
\concept{census functions}
